Cleanup
- Improve names.
- Remove redundant casts.
- Restructure nested if.
- A few other trivialities.

+++ src/main/java/primitivo/mutable/MutableByte.java
@@ -72,7 +72,7 @@ public class MutableByte implements Comparable<MutableByte> {
 	
 	@Override
 	public int hashCode() {
-		return (int) value;
+		return value;
 	}
 	
 	//@Override
